Move tabs into component, adjust query
